Add files via upload
correct do_immigration and do_ectopic_recombination bugs, output sampled immunity, number of circulating genes/strains in liver and blood separately and number of new genes generated via mutation and recombination (without checking for duplicates) and so on.

+++ src/parameters.jl
@@ -228,7 +228,7 @@ keyword constructor for the class.
 
     See description in the `immunity` field of struct `State`.
     """
-    immunity_level_max::Union{Int8, Nothing} = nothing
+    immunity_level_max::Union{Int16, Nothing} = nothing
 
     """
     Rate at which immunity is lost, per host, per gene.
